backend/ExpenseManagement/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
click==8.1.8
colorama==0.4.6
distlib==0.3.9
fastapi==0.115.6
filelock==3.17.0
greenlet==3.1.1
h11==0.14.0
idna==3.10
numpy==2.2.2
platformdirs==4.3.6
psycopg2-binary==2.9.10
pyarrow==19.0.0
pydantic==2.10.5
pydantic_core==2.27.2
pylance==0.22.0
python-dotenv==1.0.1
sniffio==1.3.1
SQLAlchemy==2.0.37
starlette==0.41.3
typing_extensions==4.12.2
uvicorn==0.34.0
virtualenv==20.29.1
